+++ code/src/cljs/sixsq/nuvla/ui/acl/spec.cljs
@@ -3,18 +3,10 @@
     [clojure.spec.alpha :as s]))
 
 
-(s/def ::users-and-groups any?)
-
-(s/def ::users-options any?)
-
-(s/def ::groups-options any?)
+(s/def ::groups any?)
 
 
-(s/def ::db (s/keys :req [::users-and-groups
-                          ::users-options
-                          ::groups-options]))
+(s/def ::db (s/keys :req [::groups]))
 
 
-(def defaults {::users-and-groups {}
-               ::users-options    []
-               ::groups-options   []})
+(def defaults {::groups []})