+++ code/src/cljs/sixsq/nuvla/ui/apps/views_detail.cljs
@@ -1118,7 +1118,7 @@
   []
   (let [tr         (subscribe [::i18n-subs/tr])
         module     (subscribe [::subs/module])
-        groups     (subscribe [::acl-subs/users-and-groups])
+        groups     (subscribe [::acl-subs/groups])
         is-vendor? (utils/is-vendor? @module)]
     (if is-vendor?
       (let [groups-from-module (utils/module->groups @module)
@@ -1176,7 +1176,7 @@
                           (when-let [user-id (and can-edit?
                                                   @(subscribe [::session-subs/active-claim]))]
                             {:owners [user-id]}))
-        ui-acl        (when acl (r/atom (acl-utils/acl->ui-acl-format acl)))]
+        ui-acl        (r/atom (when acl (acl-utils/acl->ui-acl-format acl)))]
     {:menuItem {:content (r/as-element [ShareTitle])
                 :key     "share"}
      :pane     {:key     "share-pane"